Refactor experience replay into a wrapper type
To reduce coupling, the agent must not know the implementation details
of its replay memory.
Therefore, interactions between agent and memory were previously
specified in a common trait, which allowed the agent to be generic over
the variant of replay memory used.
Unfortunately, several difficulties were encountered when trying to
unify queued and prioritized replay under this interface. This is due
to inherent differences between the two algorithms.
So now, rather than having a single agent type connecting the raw model
to a generic replay mechanism, experience replay logic has been
conceptually seperated from the agent and embedded in an optional
wrapper type instead, with a different wrapper for each replay variant.
The new design may also permit further DQN improvements to be composed
in a similar manner