Serialize plots instead of presenting them
Presenting plots visually is convenient, but it heavily restricts the
ways in which the plotted data can be used. Since images are only
intended to be viewed by the human eye, computers can have a hard time
extracting raw data from visual plots for further processing.
With serialization, the program can dump the plot as it is and let an
external tool choose what to do with the data