Fix window dimensions bug
Previously, when starting a new game, the game window's dimensions were
retrieved as soon as the window came into existence. Mysteriously,
sometimes the program crashed with a seemingly unrelated error. After
researching the bug, it turns out that the window actually starts with
a temporary set of dimensions, which is then replaced by the real one
upon entering fullscreen. This caused the wrong dimensions to be used,
which eventually led to the crash. Now, the program properly waits for
game setup to finish before it retrieves the window's dimensions