Fix window not dying inside GameInterface::end
When a game session is truncated via GameInterface::end, two things
happen: First the window is destroyed, and then the process is killed.
Window destruction is achieved by a request to the x11 server.
Previously, the x11 client buffer was accidentally not flushed after the
request, so that the window would only be destroyed once the function
had already exited.
Now the function flushes the buffer as it should and then waits for x11
to finish processing its request