Require double-buffering for ui window
The program previously accepted single-buffered fbconfigs for the ui
window. However, the ui didn't actually work for such configs, since it
relied on glXSwapBuffers flushing the gl queue, and buffer-swapping is a
no-op on single-buffered configs. While this bug can be easily fixed by
explicit calls to glFlush/glFinish, most machines nowadays support
double-buffered configs anyway, so I have decided that it is simpler to
just drop support for single-buffering