Extract color planes to match model signature
The model expects each state it receives to have the shape (8, 128, 72),
and thus each frame to have the shape (2, 128, 72) (since each state
consists of 4 frames). The outermost dimension should belong to the
color planes, of which there are 2. These planes were previously
supplied to the model in the innermost dimension, so it actually
received frames of shape (128, 72, 2). Now the planes have been
extracted to the outermost dimension, to match the signature of
(2, 128, 72)