Reorder frame dims in model signature
The model expected to receive frames of shape (128, 72), yet the frames
given were actually laid out in memory as if they had the shape
(72, 128). Thus the frame pixels looked scrambled to the model.
Remarkably, the model still managed to learn despite this mistake, but
the performance of its convolutional layers was probably harmed. Now the
model signature has been changed to expect frames of shape (72, 128)
instead