Ensure parent dirs are created before inner files
There are currently two cases where the program writes output to files:
When saving a training checkpoint (Inside the "saved" dir), or when
plotting the progress of training so far (Inside the "progress" dir).
Previously, when files were created inside a dir, the program assumed
the dir to exist but didn't actually create it. If the user didn't
create the dir beforehand, and the program tried to create files inside
it, the program would panic. Now the program creates those dirs on its
own.
NOTE: This only applies to output dirs, not to input dirs. Input dirs
are still assumed to exist (and to contain valid input)