Increase wait duration for game window starts
When starting a new game window, the program needs to wait for the
window to initialize properly. It achieves this by waiting for a fixed
duration, after which it assumes that the window is ready. The previous
duration of 2 seconds didn't always suffice, so it has now been changed
to 5 seconds