Enable n-step rewards with n = 3
Until now, n-step rewards were disabled for the agent (by setting
n = 1). I read somewhere that combining n-step with off-policy learning
is theoretically unsound, and didn't manage to find any empirical
benefit in enabling it during my testing, therefore it made sense to
leave it disabled.
After further research on the topic, I'm quite convinced that n-step may
actually be beneficial for the agent and that I've blown the theoretical
problems out of proportion. It seems to be especially helpful with large
replay buffers. I've recently managed to train the agent successfully
with n = 3, so I might as well enable n-step