Differentiate between rc and non-rc transitions
For the mono agent, it made sense to reference-count the frames of each
CompressedState, in order to save memory. For a distributed agent,
however, reference-counting is not so trivial, since pointers can't be
communicated between machines. Therefore, for now, the distributed agent
doesn't do any reference-counting. The mono agent still does it however