Select dummy UI impl for the distributed agent
In the mono agent, which runs a single instance of the environment,
enabling UI is sensible, since it allows us to see the game from the
perspective of the model, and it only causes a single extra window to
be created. The UI is not strictly necessary, but it is a nice
embellishment.
The distributed agent, in contrast, is based on the premise of running
as many actors as possible in parallel. Enabling UI for each actor means
that many UI windows will be created, leading to additional overhead.
Unfortunately, the extra overhead of maintaining so many windows might
be significant enough to force a reduction in the number of actors,
which will likely harm the agent's performance.
Since the UI isn't really necessary, we disable it for the distributed
agent by selecting its dummy impl